Initial change of the cell shapes on the 3rd tab to look more like E. Coli.

+++ simulations-java/simulations/gene-expression-basics/src/edu/colorado/phet/geneexpressionbasics/common/model/BioShapeUtils.java
@@ -329,7 +329,7 @@ public class BioShapeUtils {
      * @param rotationAngle
      * @return
      */
-    public static Shape createEColiLikeShape( Point2D center, double width, double height, double rotationAngle, int seed ) {
+    public static Shape createEColiLikeShape( Point2D center, double width, double height, double rotationAngle, long seed ) {
         assert width > height; // Param checking.  Can't create the needed shape if this isn't true.
 
         // Tweakable parameters that affect number of points used to define
